requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.12.2
bleach==6.0.0
build==0.10.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
cloudinary==1.33.0
comtypes==1.2.0
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
distlib==0.3.7
docutils==0.19
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.40.0
frozendict==2.3.8
gast==0.4.0
geojson==2.5.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
html5lib==1.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
jaraco.classes==3.2.3
keras==2.13.1
keyring==23.13.1
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.2
mailtrap==2.0.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==9.1.0
moviepy==1.0.3
multitasking==0.0.11
numpy==1.25.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.10.0
proglog==0.1.10
prompt-toolkit==1.0.14
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydub==0.25.1
pyfiglet==0.8.post1
Pygments==2.15.1
PyInquirer==1.0.3
pyowm==3.3.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
readme-renderer==37.3
regex==2023.6.3
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
rsa==4.9
six==1.16.0
soupsieve==2.4.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc2
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0rc2
termcolor==2.3.0
tkinterdnd2==0.3.0
tqdm==4.65.0
tweepy==4.14.0
twine==4.0.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.3
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
yfinance==0.2.22
zipp==3.15.0
